requirements.txt
numpy
opencv-python
pytz
Pillow
requests
qrcode

#pytest
#pytest-pylint